+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + Move login to GenericForm
 + Add payment form
 + Add payment page
++ Refactor generic page
 
 ## 0.8.0
 + Remove token middleware on logout